+++ CHANGELOG.md
@@ -35,3 +35,4 @@ fix error in Entity sensor.wi_fi_signal_quality
 ## v0.1.7
 Added `Thermostat sensor` ON or OFF
 Added `lambdaStatus`
+Added `mode` boiler operation names to status